+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Released on 14/12/2020
     - Explorer tabs have now 70% of layout height, while logging area is 30%
   - Auth page:
     - align popup text to center
+- Keybindings:
+  - `L`: Refresh directory content
 - Bugfix:
   - Fixed memory vulnerability in Windows version